+++ BrickGame/core/src/com/mictlan/brick/entities/PowerUp.java
@@ -9,7 +9,7 @@ import com.mictlan.brick.utils.ColorFactory;
 public class PowerUp extends GameObject {
     private int velY = 200;
 
-    public PowerUp(int x, int y, int width, int height ) {
+    public PowerUp(float x, float y, float width, float height ) {
         super(x, y, width, height);
         color = ColorFactory.getColor(255, 215,	0);
         hitbox = new Rectangle(x, y, width, height);
